Salirse Partido Control Exception

+++ NBP/src/controlador/GestionSalirsePartido.java
@@ -12,6 +12,7 @@ import modelo.BD;
 import modelo.Jugador;
 import vista.MenuJugador;
 import vista.SalirsePartido;
+import modelo.Partido;
 
 public class GestionSalirsePartido implements ActionListener {
 	private SalirsePartido vista = new SalirsePartido();
@@ -60,21 +61,34 @@ public class GestionSalirsePartido implements ActionListener {
 	}
 
 	private void salirPartido() {
-		String del = "Delete from Jugador_Partido where partido=\"" + vista.textField.getText() + "\" and ID_jug=\""
-				+ jugador.getCorreo() + "\"";
-		this.bd.Delete(del);
-		String sel = "SELECT Partido.id_anfitrion FROM Partido WHERE Partido.cod_partido = "
-				+ vista.textField.getText();
-		String anfitrion = bd.SelectEscalar(sel).toString();
-		if (anfitrion.equalsIgnoreCase(jugador.getCorreo())) {
-			del = "DELETE FROM Partido WHERE Partido.cod_partido = " + vista.textField.getText();
-			bd.Delete(del);
-			JOptionPane.showMessageDialog(null, "Al ser el anfitrion, el partido ha sido eliminado",
-					"Se ha salido correctamente", JOptionPane.INFORMATION_MESSAGE);
-		} else {
-			JOptionPane.showMessageDialog(null, "Ya no pertenece a dicho partido", "Se ha salido correctamente",
-					JOptionPane.INFORMATION_MESSAGE);
+		if(vista.textField.getText().equals("")) {
+			JOptionPane.showMessageDialog(null, "Por favor complete los datos del id partido", "ERROR AL SALIRSE PARTIDO!!",
+					JOptionPane.ERROR_MESSAGE);
+		}else {
+			long id = Long.parseLong(vista.textField.getText());
+			if(Partido.existePartido(id)) {
+				JOptionPane.showMessageDialog(null, "Ese id de partido no existe, por favor introduzca un id correcto", "ERROR AL SALIRSE PARTIDO!!",
+						JOptionPane.ERROR_MESSAGE);
+			}else {
+			
+				String del = "Delete from Jugador_Partido where partido=\"" + vista.textField.getText() + "\" and ID_jug=\""
+						+ jugador.getCorreo() + "\"";
+				this.bd.Delete(del);
+				String sel = "SELECT Partido.id_anfitrion FROM Partido WHERE Partido.cod_partido = "
+						+ vista.textField.getText();
+				String anfitrion = bd.SelectEscalar(sel).toString();
+				if (anfitrion.equalsIgnoreCase(jugador.getCorreo())) {
+					del = "DELETE FROM Partido WHERE Partido.cod_partido = " + vista.textField.getText();
+					bd.Delete(del);
+					JOptionPane.showMessageDialog(null, "Al ser el anfitrion, el partido ha sido eliminado",
+							"Se ha salido correctamente", JOptionPane.INFORMATION_MESSAGE);
+				} else {
+					JOptionPane.showMessageDialog(null, "Ya no pertenece a dicho partido", "Se ha salido correctamente",
+							JOptionPane.INFORMATION_MESSAGE);
+				}
+			}
 		}
+		
 	}
 
 	private void volverMP() {
